Improve: Phase 3 database and performance improvements
This commit implements Phase 3 improvements focused on performance,
data integrity, and financial accuracy:

1. Performance: Add indexes to foreign keys (3 tables)
   - expense.event_id: Added index for faster filtering
   - payment.event_id: Added index for faster filtering
   - task.event_id: Added index for faster filtering
   - Impact: Faster queries when filtering by event

2. Financial Accuracy: Float → Numeric for money fields (4 models)
   - participant: calculated_price, manual_price_override → Numeric(10, 2)
   - participant: discount_percent → Numeric(5, 2)
   - payment: amount → Numeric(10, 2)
   - expense: amount → Numeric(10, 2)
   - income: amount → Numeric(10, 2)
   - Impact: No more floating-point rounding errors
   - Impact: Correct cent-accurate calculations

3. Data Integrity: Add unique constraint (1 table)
   - setting.event_id: Added unique constraint
   - Impact: Ensures only one setting per event
   - Impact: Database-level validation

4. Database Migration
   - New Alembic migration with upgrade and downgrade
   - Supports both SQLite and PostgreSQL
   - Converts existing Float data to Numeric
   - Backward compatible

BREAKING: Money fields now use Decimal type in Python
- Services/Schemas may need adjustment
- Excel exports handle Decimal → float conversion
- Invoice generation supports Decimal formatting

Impact: Local single-user application with improved performance,
financial accuracy, and data integrity

+++ app/models/income.py
@@ -1,5 +1,5 @@
 """Income Model für zusätzliche Einnahmen (z.B. Zuschüsse)"""
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, Date, DateTime, Text
+from sqlalchemy import Column, Integer, String, Numeric, ForeignKey, Date, DateTime, Text
 from sqlalchemy.orm import relationship
 from app.database import Base
 from app.utils.datetime_helper import get_utc_timestamp
@@ -12,7 +12,7 @@ class Income(Base):
     id = Column(Integer, primary_key=True, index=True)
     event_id = Column(Integer, ForeignKey("events.id"), nullable=False, index=True)
     name = Column(String(200), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(10, 2), nullable=False)
     date = Column(Date, nullable=False)
     description = Column(Text, nullable=True)
     receipt_file_path = Column(String(500), nullable=True)  # Pfad zum hochgeladenen Beleg (PDF/Bild)
